Operation file hosts operation functions refactor
