+++ src/index.ts
@@ -1,5 +1,6 @@
 import { DurableObject } from "cloudflare:workers";
 import { createResponse, QueryRequest, QueryTransactionRequest } from './utils';
+import { enqueueOperation } from "./operation";
 
 const DURABLE_OBJECT_ID = 'sql-durable-object';
 
@@ -25,121 +26,6 @@ export class DatabaseDurableObject extends DurableObject {
         this.sql = ctx.storage.sql;
     }
 
-    executeQuery(sql: string, params?: any[]): any[] {
-        try {
-            let cursor;
-
-            if (params && params.length) {
-                cursor = this.sql.exec(sql, params);
-            } else {
-                cursor = this.sql.exec(sql);
-            }
-
-            const result = cursor.toArray();
-            return result;
-        } catch (error) {
-            console.error('SQL Execution Error:', error);
-            throw error;
-        }
-    }
-
-    async executeTransaction(queries: { sql: string; params?: any[] }[]): Promise<any[]> {
-        const results = [];
-        let transactionBookmark: any | null = null;
-
-        try {
-            // Create a storage bookmark before starting the transaction.
-            transactionBookmark = await this.ctx.storage.getCurrentBookmark();
-
-            for (const queryObj of queries) {
-                const { sql, params } = queryObj;
-                const result = this.executeQuery(sql, params);
-                results.push(result);
-            }
-
-            transactionBookmark = null;
-            return results;
-        } catch (error) {
-            console.error('Transaction Execution Error:', error);
-
-            /**
-             * If an error occurs during the transaction, we can restore the storage to the state
-             * before the transaction began by using the bookmark we created before starting the
-             * transaction.
-             */
-            if (transactionBookmark) {
-                await this.ctx.storage.onNextSessionRestoreBookmark(transactionBookmark);
-                await this.ctx.abort();
-            }
-
-            throw error;
-        }
-    }
-
-    async enqueueOperation(
-        queries: { sql: string; params?: any[] }[],
-        isTransaction: boolean
-    ): Promise<Response> {
-        const MAX_WAIT_TIME = 25000;
-        return new Promise((resolve, reject) => {
-            const timeout = setTimeout(() => {
-                reject(createResponse(undefined, 'Operation timed out.', 503));
-            }, MAX_WAIT_TIME);
-
-            this.operationQueue.push({
-                queries,
-                isTransaction,
-                resolve: (value) => {
-                    clearTimeout(timeout);
-                    resolve(value);
-                },
-                reject: (reason) => {
-                    clearTimeout(timeout);
-                    reject(reason);
-                }
-            });
-
-            this.processNextOperation().catch((err) => {
-                console.error('Error processing operation queue:', err);
-            });
-        });
-    }
-
-    async processNextOperation() {
-        if (this.processingOperation) {
-            // Already processing an operation
-            return;
-        }
-
-        if (this.operationQueue.length === 0) {
-            // No operations remaining to process
-            return;
-        }
-
-        this.processingOperation = true;
-
-        const { queries, isTransaction, resolve, reject } = this.operationQueue.shift()!;
-
-        try {
-            let result;
-
-            if (isTransaction) {
-                result = await this.executeTransaction(queries);
-            } else {
-                const { sql, params } = queries[0];
-                result = this.executeQuery(sql, params);
-            }
-
-            resolve(createResponse(result, undefined, 200));
-        } catch (error: any) {
-            console.error('Operation Execution Error:', error);
-            reject(createResponse(undefined, error.message || 'Operation failed.', 500));
-        } finally {
-            this.processingOperation = false;
-            await this.processNextOperation();
-        }
-    }
-
     async queryRoute(request: Request): Promise<Response> {
         try {
             const contentType = request.headers.get('Content-Type') || '';
@@ -162,7 +48,7 @@ export class DatabaseDurableObject extends DurableObject {
                     return { sql, params };
                 });
 
-                const response = await this.enqueueOperation(queries, true);
+                const response = await enqueueOperation(this, queries, true);
                 return response;
             } else if (typeof sql !== 'string' || !sql.trim()) {
                 return createResponse(undefined, 'Invalid or empty "sql" field.', 400);
@@ -171,7 +57,7 @@ export class DatabaseDurableObject extends DurableObject {
             }
     
             const queries = [{ sql, params }];
-            const response = await this.enqueueOperation(queries, false);
+            const response = await enqueueOperation(this, queries, false);
             return response;
         } catch (error: any) {
             console.error('Query Route Error:', error);
